I split me apart so each feature can be worked on individually
