+++ capstonFE/src/components/Me.jsx
@@ -2,6 +2,11 @@ import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import "../App.css";
+import AccountHeader from "./UserAccount/AccountHeader";
+import AccountInfo from "./UserAccount/AccountInfo";
+import ActivityTabs from "./UserAccount/ActivityTab";
+import MyActivity from "./UserAccount/MyActivity";
+import FriendsActivity from "./UserAccount/FriendsActivity";
 import Friends from "./Friends";
 
 function Me() {
@@ -13,10 +18,7 @@ function Me() {
   const [userData, setUserData] = useState(null);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState(null);
-
   const [activeTab, setActiveTab] = useState("myActivity");
-  const [editFormVisible, setEditFormVisible] = useState(false);
-
   const [friendsActivity, setFriendsActivity] = useState({
     reviews: [],
     favorites: [],
@@ -24,38 +26,15 @@ function Me() {
   const [isLoadingFriendsActivity, setIsLoadingFriendsActivity] =
     useState(false);
 
-  const [formData, setFormData] = useState({
-    username: "",
-    email: "",
-    password: "",
-  });
-
   const token = localStorage.getItem("token") || null;
 
-  useEffect(() => {}, []);
-
-  const updateSelf = async (id, username, email, password) => {
-    try {
-      const response = await axios.get(`${API_BASE_URL}/api/users/me`, {
-        headers: { Authorization: `Bearer ${token}` },
-      });
-
-      if (response.data && response.data.id) {
-        const userId = response.data.id;
-        const updateResponse = await axios.put(
-          `${API_BASE_URL}/api/users/${userId}/edit`,
-          formData,
-          {
-            headers: { Authorization: `Bearer ${token}` },
-          }
-        );
-        setFormData({ username: "", email: "", password: "" });
-      }
-    } catch (error) {
-      console.error("Error updating user:", error);
-      return null;
+  useEffect(() => {
+    if (!token) {
+      navigate("/login");
+      return;
     }
-  };
+    userUpdate();
+  }, []);
 
   const userUpdate = async () => {
     try {
@@ -71,7 +50,6 @@ function Me() {
         );
 
         const reviews = reviewsResponse.data || [];
-
         const enhancedReviews = reviews.map((review) => ({
           ...review,
           albumName: review.headline,
@@ -140,14 +118,6 @@ function Me() {
   };
 
   useEffect(() => {
-    if (!token) {
-      navigate("/login");
-      return;
-    }
-    userUpdate();
-  }, []);
-
-  useEffect(() => {
     if (activeTab === "friendsActivity" && userData && userData.id) {
       fetchFriendsActivity(userData.id);
     }
@@ -157,30 +127,6 @@ function Me() {
     localStorage.removeItem("token");
     localStorage.removeItem("user");
     navigate("/login");
-  };
-
-  // show/hide edit form
-  const handleFormShow = () => {
-    setEditFormVisible(!editFormVisible);
-  };
-
-  const handleChange = (event) => {
-    setFormData({ ...formData, [event.target.name]: event.target.value });
-  };
-
-  // submit profile changes
-  const handleEdit = async (event) => {
-    event.preventDefault();
-    if (!formData.username || !formData.email || formData.password === "") {
-      alert("Please fill out all fields.");
-      return;
-    }
-    await updateSelf(
-      userData.id,
-      formData.username,
-      formData.email,
-      formData.password
-    );
   };
 
   if (isLoading) {
@@ -203,277 +149,21 @@ function Me() {
   return (
     <div className="account-page">
       <div className="account-container">
-        <div className="account-header">
-          <h1>My Account</h1>
-          <button className="logout-button" onClick={handleLogout}>
-            Log Out
-          </button>
-        </div>
-
+        <AccountHeader onLogout={handleLogout} />
         <div className="account-content">
           {userData ? (
-            <div className="user-info">
-              <h2>Account Information</h2>
-              <div className="info-item">
-                <div className="userData">
-                  <span className="label">Username:</span>
-                  <span className="value">
-                    {userData.username || "Not available"}
-                  </span>
-                </div>
-
-                <div className="editWrapper">
-                  <button className="edit-button" onClick={handleFormShow}>
-                    {editFormVisible ? "Cancel" : "Edit Profile"}
-                  </button>
-                  {editFormVisible && (
-                    <form onSubmit={handleEdit}>
-                      <label>
-                        Username:
-                        <input
-                          type="text"
-                          name="username"
-                          onChange={handleChange}
-                          value={formData?.username}
-                          required
-                        />
-                      </label>
-                      <label>
-                        Email:
-                        <input
-                          type="email"
-                          name="email"
-                          onChange={handleChange}
-                          value={formData?.email}
-                          required
-                        />
-                      </label>
-                      <label htmlFor="password">
-                        Password:
-                        <input
-                          type="text"
-                          onChange={handleChange}
-                          name="password"
-                          value={formData?.password}
-                          required
-                        />
-                      </label>
-                      <button className="edit-button" type="submit">
-                        Save Changes
-                      </button>
-                    </form>
-                  )}
-                </div>
-              </div>
-              {userData.email && (
-                <div className="info-item">
-                  <span className="label">Email:</span>
-                  <span className="value">{userData.email}</span>
-                </div>
+            <>
+              <AccountInfo userData={userData} />
+              <ActivityTabs activeTab={activeTab} setActiveTab={setActiveTab} />
+              {activeTab === "myActivity" && <MyActivity userData={userData} />}
+              {activeTab === "friendsActivity" && (
+                <FriendsActivity
+                  friendsActivity={friendsActivity}
+                  isLoading={isLoadingFriendsActivity}
+                />
               )}
-
-              <div className="activity-tabs">
-                <button
-                  className={`tab-button ${
-                    activeTab === "myActivity" ? "active" : ""
-                  }`}
-                  onClick={() => setActiveTab("myActivity")}>
-                  My Activity
-                </button>
-                <button
-                  className={`tab-button ${
-                    activeTab === "friendsActivity" ? "active" : ""
-                  }`}
-                  onClick={() => setActiveTab("friendsActivity")}>
-                  Friends' Activity
-                </button>
-                <button
-                  className={`tab-button ${
-                    activeTab === "friends" ? "active" : ""
-                  }`}
-                  onClick={() => setActiveTab("friends")}>
-                  See and Find Friends
-                </button>
-              </div>
-
-              {activeTab === "myActivity" ? (
-                <div className="account-activity">
-                  <h2>Your Activity</h2>
-
-                  {/* Favorite Albums */}
-                  {userData.favorites &&
-                  Array.isArray(userData.favorites) &&
-                  userData.favorites.length > 0 ? (
-                    <div className="favorites-section">
-                      <h3>Favorite Albums</h3>
-                      <ul className="favorites-list">
-                        {userData.favorites.map((album, index) => (
-                          <li key={index}>
-                            <strong>{album.name || "Unknown Album"}</strong>
-                            <p>Artist: {album.artist || "Unknown Artist"}</p>
-                            {album.image && (
-                              <img
-                                src={album.image}
-                                alt={album.name || "Album Cover"}
-                                style={{
-                                  width: "100px",
-                                  borderRadius: "8px",
-                                  marginTop: "5px",
-                                  cursor: "pointer",
-                                }}
-                                onClick={() =>
-                                  navigate(`/album/${album.spotify_id}`)
-                                }
-                              />
-                            )}
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  ) : (
-                    <p>No favorite albums yet.</p>
-                  )}
-
-                  {/* User Reviews */}
-                  {userData.reviews &&
-                  Array.isArray(userData.reviews) &&
-                  userData.reviews.length > 0 ? (
-                    <div className="reviews-section">
-                      <h3>Recent Reviews</h3>
-                      <ul className="reviews-list">
-                        {userData.reviews.map((review, index) => (
-                          <li key={index}>
-                            <img
-                              src={review.album_image}
-                              alt={review.album_name || "Album Cover"}
-                              style={{
-                                width: "100px",
-                                borderRadius: "8px",
-                                marginTop: "5px",
-                                cursor: "pointer",
-                              }}
-                              onClick={() =>
-                                navigate(`/album/${review.album_spotify_id}`)
-                              }
-                            />
-                            <p>
-                              <strong>
-                                {review.album_name ||
-                                  review.album?.name ||
-                                  "Unknown Album"}{" "}
-                              </strong>
-                            </p>
-                            <span> Rating: {review.rating}/5</span>
-                            <p>{review.review || "No review text provided."}</p>
-                            <small>By: {review.username}</small>
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  ) : (
-                    <p>No reviews yet.</p>
-                  )}
-                </div>
-              ) : activeTab === "friendsActivity" ? (
-                <div className="friends-activity">
-                  <h2>Friends' Activity</h2>
-
-                  {isLoadingFriendsActivity ? (
-                    <div className="loading">Loading friends' activity...</div>
-                  ) : (
-                    <>
-                      {/* Friends' Favorite Albums */}
-                      {friendsActivity.favorites &&
-                      friendsActivity.favorites.length > 0 ? (
-                        <div className="favorites-section">
-                          <h3>Friends' Favorite Albums</h3>
-                          <ul className="favorites-list">
-                            {friendsActivity.favorites.map(
-                              (favorite, index) => (
-                                <li key={index}>
-                                  <strong>
-                                    {favorite.name || "Unknown Album"}
-                                  </strong>
-                                  <p>
-                                    Artist:{" "}
-                                    {favorite.artist || "Unknown Artist"}
-                                  </p>
-                                  <small>Liked by: {favorite.username}</small>
-                                  {favorite.image && (
-                                    <img
-                                      src={favorite.image}
-                                      alt={favorite.name || "Album Cover"}
-                                      style={{
-                                        width: "100px",
-                                        borderRadius: "8px",
-                                        marginTop: "5px",
-                                        cursor: "pointer",
-                                      }}
-                                      onClick={() =>
-                                        navigate(
-                                          `/album/${favorite.spotify_id}`
-                                        )
-                                      }
-                                    />
-                                  )}
-                                </li>
-                              )
-                            )}
-                          </ul>
-                        </div>
-                      ) : (
-                        <p>No favorite albums from friends yet.</p>
-                      )}
-
-                      {/* Friends' Reviews */}
-                      {friendsActivity.reviews &&
-                      friendsActivity.reviews.length > 0 ? (
-                        <div className="reviews-section">
-                          <h3>Friends' Recent Reviews</h3>
-                          <ul className="reviews-list">
-                            {friendsActivity.reviews.map((review, index) => (
-                              <li key={index}>
-                                <img
-                                  src={review.album_image}
-                                  alt={review.album_name || "Album Cover"}
-                                  style={{
-                                    width: "100px",
-                                    borderRadius: "8px",
-                                    marginTop: "5px",
-                                    cursor: "pointer",
-                                  }}
-                                  onClick={() =>
-                                    navigate(
-                                      `/album/${review.album_spotify_id}`
-                                    )
-                                  }
-                                />
-                                <p>
-                                  <strong>
-                                    {review.album_name || "Unknown Album"}
-                                  </strong>
-                                </p>
-                                <span> Rating: {review.rating}/5</span>
-                                <p>
-                                  {review.review || "No review text provided."}
-                                </p>
-                                <small>By: {review.username}</small>
-                              </li>
-                            ))}
-                          </ul>
-                        </div>
-                      ) : (
-                        <p>No reviews from friends yet.</p>
-                      )}
-                    </>
-                  )}
-                </div>
-              ) : (
-                <div className="friends">
-                  <Friends />
-                </div>
-              )}
-            </div>
+              {activeTab === "friends" && <Friends />}
+            </>
           ) : (
             <p>No user data available.</p>
           )}
